modified:   src/main/java/Average.java

+++ src/main/java/Average.java
@@ -18,7 +18,7 @@ public class Average {
      * @return the average of all the values of nums.
      */
     public double getAvg(int[] nums){
-        int a = 0;
+        double a = 0;
         for(int i = 0; i < nums.length; i++){
             a = a + nums[i];
         }
